llm_tests/pyproject.toml
[tool.poetry]
name = "llm_tests"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
protobuf = "^3.20.0"
torch = "^1.12.0"
torchvision = "^0.13.0"
transformers = "^4.20.1"
typer = "^0.6.1"
sentencepiece = "^0.1.96"
datasets = "^2.3.2"
Pillow = "^9.2.0"
pytesseract = "^0.3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test1 = "llm_tests.test1:main"
testqa = "llm_tests.testqa:main"
